Create a macro for contact information
The blocks of contect information contain a large amount of repetitive
markup. Because this is a hastle to maintain (and somewhat error prone),
I've implemented the rendering as a Jinja macro. There's probably more
scripting going on than appropriate, but it reduces duplicated markup
significantly.

Any attributes that are inferred from the type of contact info (as in
the case of, for example, discerning LinkedIn URL from the ID) may be
explicitly overridden when the macro is invoked.